MHV-62798 MHV-62899 Fixed existing tests
